framework/requirements.txt
flask>=3.0
flask-limiter>=3.0
anthropic>=0.40.0
python-dotenv>=1.0.0
requests>=2.31.0
feedparser>=6.0.0
reportlab>=4.0
